feat(widegt): refactored and moved code for execute last mile tx to a hook
