+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@anyalt/widget",
-  "version": "1.0.10",
+  "version": "1.0.10-canary.1",
   "description": "Anyalt Widget",
   "private": false,
   "files": [
